requirements.txt
google-cloud-vision
pymupdf
python-dotenv
pillow
pytesseract
numpy
